+++ src/cell/mechanics.rs
@@ -12,49 +12,49 @@ use crate::NVERTS;
 use crate::math::geometry::calc_poly_area;
 use crate::math::capped_linear_function;
 
-pub fn calc_edge_vecs(vertex_coords: &[P2D; NVERTS as usize]) -> [P2D; NVERTS as usize] {
-    let mut r = [P2D::default(); NVERTS as usize];
-    (0..NVERTS as usize).for_each(|i| {
-        let plus_i = circ_ix_plus(i, NVERTS as usize);
+pub fn calc_edge_vecs(vertex_coords: &[P2D; NVERTS]) -> [P2D; NVERTS] {
+    let mut r = [P2D::default(); NVERTS];
+    (0..NVERTS).for_each(|i| {
+        let plus_i = circ_ix_plus(i, NVERTS);
         r[i] = vertex_coords[plus_i] - vertex_coords[i];
     });
     r
 }
 
 pub fn calc_edge_forces(
-    edge_strains: &[f32; NVERTS as usize],
-    edge_unit_vecs: &[P2D; NVERTS as usize],
+    edge_strains: &[f32; NVERTS],
+    edge_unit_vecs: &[P2D; NVERTS],
     stiffness_edge: f32,
-) -> [P2D; NVERTS as usize] {
-    let mut r = [P2D::default(); NVERTS as usize];
-    (0..NVERTS as usize).for_each(|i| r[i] = edge_strains[i] * stiffness_edge * edge_unit_vecs[i]);
+) -> [P2D; NVERTS] {
+    let mut r = [P2D::default(); NVERTS];
+    (0..NVERTS).for_each(|i| r[i] = edge_strains[i] * stiffness_edge * edge_unit_vecs[i]);
     r
 }
 
 pub fn calc_cyto_forces(
-    vertex_coords: &[P2D; NVERTS as usize],
-    unit_inward_vecs: &[P2D; NVERTS as usize],
+    vertex_coords: &[P2D; NVERTS],
+    unit_inward_vecs: &[P2D; NVERTS],
     rest_area: f32,
     stiffness_cyto: f32,
-) -> [P2D; NVERTS as usize] {
+) -> [P2D; NVERTS] {
     let area = calc_poly_area(vertex_coords);
     let areal_strain = (area / rest_area) - 1.0;
     let mag = stiffness_cyto * areal_strain;
-    let mut r = [P2D::default(); NVERTS as usize];
-    (0..NVERTS as usize).for_each(|i| r[i] = mag * unit_inward_vecs[i]);
+    let mut r = [P2D::default(); NVERTS];
+    (0..NVERTS).for_each(|i| r[i] = mag * unit_inward_vecs[i]);
     r
 }
 
 pub fn calc_rgtp_forces(
-    rac_acts: &[f32; NVERTS as usize],
-    rho_acts: &[f32; NVERTS as usize],
-    unit_inward_vecs: &[P2D; NVERTS as usize],
+    rac_acts: &[f32; NVERTS],
+    rho_acts: &[f32; NVERTS],
+    unit_inward_vecs: &[P2D; NVERTS],
     halfmax_vertex_rgtp_act: f32,
     const_protrusive: f32,
     const_retractive: f32,
-) -> [P2D; NVERTS as usize] {
+) -> [P2D; NVERTS] {
     let nvs = unit_inward_vecs.len();
-    let mut r = [P2D::default(); NVERTS as usize];
+    let mut r = [P2D::default(); NVERTS];
     for i in 0..nvs {
         let uiv = unit_inward_vecs[i];
         let ra = rac_acts[i];
